__str__() functions and Workspace.merge()
I added a __str__() function to the classes for features, syntactic objects, workspaces, and lexical arrays. I also fixed Stage.merge() and SyntacticObject.merge() so that they work correctly for external merge. Internal merge still doesn't work because Workspace.find() is not recursive. (I also moved the Feature classes to the beginning and added some stuff to the test fragment.)